Create Project Info
Use for cursor by pasting in python window in arcpro with all layers that will be used. this will then return a json file called "project info" into the project folder in file explorer. when you open that file it will open in cursor - copy and paste what it shows into cursor chat to show cursor the data you are working wit.